Add files via upload
The design of simulation of fire spread in green areas based on a map, taking into account variables such as humidity, wind, terrain. The terrain is represented by a grid of cells with different states: green, fire, water, road, rock. The simulation supports mechanisms such as fire suppression, fire addition, road creation and wind influence. The user can modify parameters in real time, and the results are displayed as animations.